requirements.txt
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0 